+++ src/Fabricator.Core/Environment/ISystemPlatform.cs
@@ -3,4 +3,8 @@ namespace Fabricator.Core.Environment;
 public interface ISystemPlatform
 {
     bool IsMacOS { get; }
+
+    bool IsWindows { get; }
+
+    bool IsLinux { get; }
 }